+++ CHANGES.rst
@@ -25,6 +25,7 @@
 * SelectField now supports optgroup grouping
 * Comparison validators to complement ``EqualTo`` with ``>``, ``>=``, ``<``, ``<=``, ``!=``
 * DateTimeField's timezone parameter now defaults to user timezone, not UTC
+* New form filters to replace the (overloaded) use of validators to filter
 
 0.3.0
 -----